requirements.txt
django
djangorestframework
ultralytics
opencv-python-headless
gunicorn
whitenoise
